Refactoring Class model and program classes
