+++ prove/Develop03/Program.cs
@@ -1,21 +1,30 @@
 using System;
 using System.Data;
 using System.Runtime.CompilerServices;
+using System.Text.RegularExpressions;
+
 
 class Program
 {
     private static string _userInput;
 
-    private static Reference _reference;
+    private static Scripture _memorizeScripture;
 
     private static bool _stillPlaying;
 
+    private static List<Scripture> _scriptures = new List<Scripture>();
+
     static void Main(string[] args)
     {
+        File _file = new File();
+
+        _scriptures = _file.GetScripturesList();
     
         _userInput = "";
 
         _stillPlaying = true;
+
+        List<Scripture> searchResult; 
 
         while (!_userInput.Equals("quit"))
         {
@@ -31,8 +40,8 @@ class Program
             switch (_userInput){
 
                 case "1":
-                    _reference = new Reference(AppData.RandomScripture());
-                    Memorize(_reference);
+                    _memorizeScripture = RandomScripture();
+                    Memorize(_memorizeScripture);
                 break;
 
                 case "2":
@@ -45,11 +54,19 @@ class Program
                     Console.Write("\nVerses: (1 or 1-7): ");
                     string verses = Console.ReadLine();
 
-                    _reference = new Reference(AppData.SpecificScripture(book, chapter, verses));
+                    searchResult = SearchScripture(book, chapter, verses);
 
-                    if (_reference != null)
+                    if (searchResult.Count == 1)
                     {
-                        Memorize(_reference);
+                        _memorizeScripture = searchResult[0];
+                    }
+                    else{
+                        _memorizeScripture = new Scripture(searchResult);
+                    } 
+
+                    if (!String.IsNullOrEmpty(_memorizeScripture.ToString()))
+                    {
+                        Memorize(_memorizeScripture);
                     }
                 break;
 
@@ -59,11 +76,19 @@ class Program
                     
                     _userInput = Console.ReadLine();
 
-                    _reference = new Reference(AppData.EvaluateScriptureExpression(_userInput));
+                    searchResult = EvaluateScriptureExpression(_userInput);
 
-                    if (_reference != null)
+                    if (searchResult.Count == 1)
                     {
-                        Memorize(_reference);
+                        _memorizeScripture = searchResult[0];
+                    }
+                    else{
+                        _memorizeScripture = new Scripture(searchResult);
+                    } 
+
+                    if (!String.IsNullOrEmpty(_memorizeScripture.ToString()))
+                    {
+                        Memorize(_memorizeScripture);
                     }
                 break;
 
@@ -75,15 +100,15 @@ class Program
         }
     }
 
-    private static string DisplayReference()
-    //Prints the reference and its associated Words list
+    private static string DisplayScripture()
+    //Prints the scripture, refence and Words list
     {
-        return $"{_reference.GetReference()}\n{_reference.GetWords()}";
+        return $"{_memorizeScripture.GetReference()}\n{_memorizeScripture.GetWords()}";
     }
 
-    private static void Memorize(Reference reference)
-    /*loops user input to call HideRandomWords method from _reference each time the user return an empty input;
-      Calls VerifyHiddenWords from _reference when the user return a value;
+    private static void Memorize(Scripture scripture)
+    /*loops user input to call HideRandomWords method from _memorizeScripture each time the user return an empty input;
+      Calls VerifyHiddenWords from _memorizeScripture when the user return a value;
       Checks for a "quit" input, ending the application. Also, will end the application if the user return an empty input
       when there's no more words to be hidden.*/
     {
@@ -92,20 +117,21 @@ class Program
         while (!input.Equals("quit"))
         {
             Console.Clear();
-            Console.WriteLine($"{DisplayReference()}");
+            Console.WriteLine($"{DisplayScripture()}");
             Console.Write("> ");
             input = Console.ReadLine();
 
             //empty input with lefting words to be hidden
             if (input.Equals("") && _stillPlaying)
             {
-                _stillPlaying = reference.HideRandomWords();
+                _stillPlaying = scripture.HideRandomWords();
             }
 
             //not empty input, calls reference for verifying the input for any match
             else if (!input.Equals(""))
             {
-                _stillPlaying = reference.VerifyHiddenWords(input);
+                _stillPlaying = true;
+                scripture.VerifyHiddenWords(input);
             }
             //empty input without anymore words to hide, will terminate the program
             else if (input.Equals("") && !_stillPlaying)
@@ -114,5 +140,120 @@ class Program
             }
         }
         System.Environment.Exit(0);
+    }
+
+    private static List<Scripture> SearchScripture(string book, string chapter, string verses)
+    //finds the scriptures that match the parameters and returns a list with one or more scriptures
+    {
+        string[] splitVerses = verses.Split("-");
+
+        //temporary list for holding user's scriptures selection
+        List<Scripture>selectedScriptures = new List<Scripture> ();
+        try{
+            //Treating compound verse scenario
+            if(verses.Count() > 1 && int.Parse(splitVerses[0]) != int.Parse(splitVerses[1]))
+            {   
+                for (int i = int.Parse(splitVerses[0]); i <= int.Parse(splitVerses[1]); i++)
+                {
+                    //Finding user informed scripture
+                    Scripture scripture = _scriptures.Find(scripture => scripture.GetBook().ToUpper() == book.ToUpper() && 
+                                                                        scripture.GetChapter() == int.Parse(chapter) &&
+                                                                        scripture.GetVerse() == i);
+
+                    //adding found scripture into selectio array
+                    selectedScriptures.Add(scripture);
+                    
+                }
+            }
+            //Treating single verse scripture scenario
+            else{
+                //finding user desired scripture
+                Scripture scripture = _scriptures.Find(scripture => scripture.GetBook().ToUpper() == book.ToUpper() && 
+                                                                        scripture.GetChapter() == int.Parse(chapter) &&
+                                                                        scripture.GetVerse() == int.Parse(splitVerses[0]));
+                
+                //adding found scripture into selectio array
+                selectedScriptures.Add(scripture);
+            }
+        }
+        catch (IndexOutOfRangeException)
+        {
+            Console.WriteLine("An error has ocurred!\nPress enter to continue");
+            Console.ReadLine();
+        }
+        return selectedScriptures;
+    }
+
+    public static Scripture RandomScripture()
+    //Picks a random index in _scriptures array and return a new instance of Reference with index associated scripture
+    {
+        Random random = new Random();
+        
+        //picking a random index from scriptures list
+        int index = random.Next(0, _scriptures.Count);
+        
+        //instantiating a new reference object with scripture from random index
+        return _scriptures[index];
+    }
+
+    public static List<Scripture> EvaluateScriptureExpression(string userInput)
+    //Evaluates user input as an regular expression, and returns a list containing user's scriptures
+    {
+        MatchCollection matches;
+
+        string pattern;
+
+        string GetBook(string userInput)
+        {
+            pattern = @"((\d*) (\w*))|(\w*)";
+
+            matches = Regex.Matches(userInput, pattern);
+
+            return matches[0].Groups[0].Value;
+        }
+
+        string GetChapter(string userInput)
+        //returns 2nd position from regex matches since it corresponds to the 1st group of matches' pattern 
+        {
+            pattern = @" (\d*):";
+
+            matches = Regex.Matches(userInput, pattern);
+
+            return matches[0].Groups[1].Value;
+        }
+
+        string GetVerses(string userInput)
+        {
+
+            pattern = @":(\d*)-(\d*)|:(\d*)";
+
+            matches = Regex.Matches(userInput, pattern);
+
+            //Regex matched for one single verse, 3 group of match
+            //returns 4th index because the the 1st contains the expression
+            if (matches[0].Groups[1].Value.Equals(""))
+            {
+                return matches[0].Groups[3].Value;
+            }
+            //Regex matched first pattern, multiple verses, again index o holds the entire expression
+            //concatenating 2nd and 3rd groups from match, which corresponds to initial and final versicles
+            else
+            {
+                string verses = "";
+                verses = matches[0].Groups[1].Value;
+
+                verses +="-"+matches[0].Groups[2].Value;
+
+                return verses;
+            }            
+        }
+
+        //returning the result from SpecificScripture Method
+
+        string book = GetBook(userInput);
+        string chapter = GetChapter(userInput);
+        string verses = GetVerses(userInput);
+
+        return SearchScripture(book, chapter, verses);
     }
 }
