+++ prove/Develop03/Word.cs
@@ -3,19 +3,19 @@ using System.Collections.Specialized;
 public class Word
 {
     private string _word;
-    private bool _isVisible;
+    private bool _isHidden;
 
     public Word (string word)
     {
         _word = word;
-        _isVisible = true;
+        _isHidden = false;
     }
 
     public string GetWord()
     //returns _word real value if _isVisible is defined as true, or returns a string of "_" 
     //with the same length from _word
     {
-        if (_isVisible)
+        if (!_isHidden)
         {
             return _word;
         }
@@ -25,28 +25,28 @@ public class Word
         }
     }
 
-    public bool GetVisibility()
+    public bool IsHidden()
     {
-        return _isVisible;
+        return _isHidden;
     }
 
     public void SetHidden()
     {
-        _isVisible = false;
+        _isHidden = true;
     }
 
-    public void SetVisible()
+    public void UnSetHidden()
     {
-        _isVisible = true;
+        _isHidden = false;
     }
 
     public bool CompareWord(string input)
-    //compares _word with input parameterm, and calls SetVisible if there's a match.
+    //compares _word with input parameterm, and calls UnSetHidden if there's a match.
     {
         bool updated = false;
         if (_word.ToUpper().Equals(input.ToUpper()))
         {
-            SetVisible();
+            UnSetHidden();
             updated = true;
         }
         return updated;
